Some changes to how I am developing
I am going to write Unit tests and run them after every commit on CircleCI, I have written a docker environment for development. Rough setup of a theme system, based on a config file. Will try to do the same for uncompiled assets and gitignore the compiled assets. Getting back to working on this alot. More to come.